Fix GHA jobs not being able to build new R versions
* Remove redundant get_r_versions script
* Use CRAN's list of R versions as the only needed source of truth
* Print build matrix for debugging help
* Add all-patch version specifier for full distro rebuilds
